method_three/requirements.txt
# Core web scraping and automation
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
seleniumbase>=4.20.0

# Firebase integration
firebase-admin>=7.1.0

# Google API integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# UI automation and input control
pyautogui>=0.9.54
pynput>=1.7.6

# Data processing
pandas>=2.1.0

# Additional utilities
python-dateutil>=2.8.2
